+++ src/routes/(v2)/Navigation.svelte
@@ -18,6 +18,7 @@
 		<li><a href="/v2/issue-309-unions">Schema union</a></li>
 		<li><a href="/v2/on-change-event">onChange event</a></li>
 		<li><a href="/v2/defaults-fail">Nested defaults</a></li>
+		<li><a href="/v2/validators-clear">Clear validators</a></li>
 	</ul>
 </nav>