+++ src/routes/(v2)/v2/validators-clear/schema.ts
@@ -1,0 +1,5 @@
+import { z } from 'zod';
+
+export const schema = z.object({
+	name: z.string().min(5)
+});